# Tweaks to reduce number of allocations in `regal lint` hot path