--- a/topdown/arithmetic.go
+++ b/topdown/arithmetic.go
@@ -67,7 +67,7 @@ func builtinPlus(_ BuiltinContext, operands []*ast.Term, iter func(*ast.Term) er
 	y, ok2 := n2.Int()
 
 	if ok1 && ok2 && inSmallIntRange(x) && inSmallIntRange(y) {
-		return iter(ast.IntNumberTerm(x + y))
+		return iter(ast.InternedIntNumberTerm(x + y))
 	}
 
 	f, err := arithPlus(builtins.NumberToFloat(n1), builtins.NumberToFloat(n2))
@@ -91,7 +91,7 @@ func builtinMultiply(_ BuiltinContext, operands []*ast.Term, iter func(*ast.Term
 	y, ok2 := n2.Int()
 
 	if ok1 && ok2 && inSmallIntRange(x) && inSmallIntRange(y) {
-		return iter(ast.IntNumberTerm(x * y))
+		return iter(ast.InternedIntNumberTerm(x * y))
 	}
 
 	f, err := arithMultiply(builtins.NumberToFloat(n1), builtins.NumberToFloat(n2))
@@ -171,7 +171,7 @@ func builtinMinus(_ BuiltinContext, operands []*ast.Term, iter func(*ast.Term) e
 		y, oky := n2.Int()
 
 		if okx && oky && inSmallIntRange(x) && inSmallIntRange(y) {
-			return iter(ast.IntNumberTerm(x - y))
+			return iter(ast.InternedIntNumberTerm(x - y))
 		}
 
 		f, err := arithMinus(builtins.NumberToFloat(n1), builtins.NumberToFloat(n2))
@@ -213,7 +213,7 @@ func builtinRem(_ BuiltinContext, operands []*ast.Term, iter func(*ast.Term) err
 				return fmt.Errorf("modulo by zero")
 			}
 
-			return iter(ast.IntNumberTerm(x % y))
+			return iter(ast.InternedIntNumberTerm(x % y))
 		}
 
 		op1, err1 := builtins.NumberToInt(n1)