--- a/topdown/graphql.go
+++ b/topdown/graphql.go
@@ -295,7 +295,7 @@ func builtinGraphQLParseAndVerify(_ BuiltinContext, operands []*ast.Term, iter f
 	var err error
 
 	unverified := ast.ArrayTerm(
-		ast.BooleanTerm(false),
+		ast.InternedBooleanTerm(false),
 		ast.NewTerm(ast.NewObject()),
 		ast.NewTerm(ast.NewObject()),
 	)
@@ -353,7 +353,7 @@ func builtinGraphQLParseAndVerify(_ BuiltinContext, operands []*ast.Term, iter f
 
 	// Construct return value.
 	verified := ast.ArrayTerm(
-		ast.BooleanTerm(true),
+		ast.InternedBooleanTerm(true),
 		ast.NewTerm(queryResult),
 		ast.NewTerm(querySchema),
 	)
@@ -421,10 +421,10 @@ func builtinGraphQLIsValid(_ BuiltinContext, operands []*ast.Term, iter func(*as
 		queryDoc, err = objectToQueryDocument(x)
 	default:
 		// Error if wrong type.
-		return iter(ast.BooleanTerm(false))
+		return iter(ast.InternedBooleanTerm(false))
 	}
 	if err != nil {
-		return iter(ast.BooleanTerm(false))
+		return iter(ast.InternedBooleanTerm(false))
 	}
 
 	switch x := operands[1].Value.(type) {
@@ -434,23 +434,23 @@ func builtinGraphQLIsValid(_ BuiltinContext, operands []*ast.Term, iter func(*as
 		schemaDoc, err = objectToSchemaDocument(x)
 	default:
 		// Error if wrong type.
-		return iter(ast.BooleanTerm(false))
+		return iter(ast.InternedBooleanTerm(false))
 	}
 	if err != nil {
-		return iter(ast.BooleanTerm(false))
+		return iter(ast.InternedBooleanTerm(false))
 	}
 
 	// Validate the query against the schema, erroring if there's an issue.
 	schema, err := convertSchema(schemaDoc)
 	if err != nil {
-		return iter(ast.BooleanTerm(false))
+		return iter(ast.InternedBooleanTerm(false))
 	}
 	if err := validateQuery(schema, queryDoc); err != nil {
-		return iter(ast.BooleanTerm(false))
+		return iter(ast.InternedBooleanTerm(false))
 	}
 
 	// If we got this far, the GraphQL query passed validation.
-	return iter(ast.BooleanTerm(true))
+	return iter(ast.InternedBooleanTerm(true))
 }
 
 func builtinGraphQLSchemaIsValid(_ BuiltinContext, operands []*ast.Term, iter func(*ast.Term) error) error {
@@ -464,15 +464,15 @@ func builtinGraphQLSchemaIsValid(_ BuiltinContext, operands []*ast.Term, iter fu
 		schemaDoc, err = objectToSchemaDocument(x)
 	default:
 		// Error if wrong type.
-		return iter(ast.BooleanTerm(false))
+		return iter(ast.InternedBooleanTerm(false))
 	}
 	if err != nil {
-		return iter(ast.BooleanTerm(false))
+		return iter(ast.InternedBooleanTerm(false))
 	}
 
 	// Validate the schema, this determines the result
 	_, err = convertSchema(schemaDoc)
-	return iter(ast.BooleanTerm(err == nil))
+	return iter(ast.InternedBooleanTerm(err == nil))
 }
 
 func init() {