Concluding my quest to reduce the number of allocations in the hot path for `regal lint` for this time around. This PR mainly does so by reusing pointers to boolean and integer terms where these are determined not to be mutated later.

The result is another ~4 million allocations reduced when linting Regal against its own bundle. These improvements should however help reduce allocations in pretty much any evaluation.

**opa main**
```
BenchmarkRegalLintingItself-10    1	3195257584 ns/op	6496097784 B/op	120108808 allocs/op
```

**PR branch**
```
BenchmarkRegalLintingItself-10    1	3132126333 ns/op	6376318224 B/op	116163318 allocs/op
```