@srenatus 2024-11-13
Ran with benchstat, just for completeness...
```
goos: darwin
goarch: amd64
pkg: github.com/styrainc/regal/pkg/linter
cpu: Intel(R) Core(TM) i9-9880H CPU @ 2.30GHz
                      │ main.bench  │      less-allocs.bench      │
                      │   sec/op    │   sec/op    vs base         │
RegalLintingItself-16   3.797 ± 10%   3.557 ± 8%  ~ (p=0.240 n=6)

                      │  main.bench  │         less-allocs.bench          │
                      │     B/op     │     B/op      vs base              │
RegalLintingItself-16   6.053Gi ± 0%   5.943Gi ± 0%  -1.83% (p=0.002 n=6)

                      │ main.bench  │         less-allocs.bench         │
                      │  allocs/op  │  allocs/op   vs base              │
RegalLintingItself-16   120.1M ± 0%   116.2M ± 0%  -3.28% (p=0.002 n=6)
```