@anderseknert 2024-11-13
@srenatus existing functions are sometimes used like this: https://github.com/open-policy-agent/opa/blob/c47b8c1723fad71e525ce4c1e2b31c1c39aa6e60/ast/parser_ext.go#L286

i.e mutating the pointer returned. If we want to have a single function to create bool/int terms, we'd have to at least identify all uses where mutation can happen after the pointer is returned. Having separate functions for cached values seemed like at least some way to try and communicate intent.