@srenatus 2024-11-13
Ah right, these are _terms_! Yeah that makes sense, many of them will have a location. OK I'm convinced. I'm little on the fence about the "cached" terminology... a cache somehow carries the connotation that you can have a hit or a miss, or that something is cached for a while and then isn't. Would "Static" seem like a valid alternative? "Interned"? 🤔