@anderseknert 2024-11-13
Interned sounds good to me! I'll make that change.